requirements.txt
pyyaml
glom